apps/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
pillow>=10.0.0
numpy>=1.26.0
rawpy>=0.21.0
